move code around into a structure I like more
