+++ webpack/components/controls/SliderInput.js
