+++ webpack/components/shapes/FlowerOfLife.js
