+++ webpack/components/shapes/Square.js
