+++ webpack/containers/CircleConfig.js
@@ -2,7 +2,7 @@ import React  from 'react';
 import { connect } from 'react-redux';
 import { CIRCLE_CONFIG } from '../shapeConstants';
 import { changeRadius } from '../actions/changeCircleConfig';
-import CircleForm from "../components/CircleForm";
+import CircleForm from "../components/forms/CircleForm";
 
 const mapStateToProps = state => ({
     radius: state[CIRCLE_CONFIG].radius,
