+++ webpack/containers/SquareConfig.js
@@ -3,7 +3,7 @@ import { SQUARE_CONFIG } from '../shapeConstants';
 import { connect } from 'react-redux';
 
 import { changeSideLength } from '../actions/changeSquareConfig';
-import SquareForm from "../components/SquareForm";
+import SquareForm from "../components/forms/SquareForm";
 
 const mapStateToProps = state => ({
     length: state[SQUARE_CONFIG].length,
